Node resolver: support IPv6 service IPs
This is a quick fix to support IPv6 service IPs when processing services
for /etc/hosts on the node. If a service exists but no A record is found,
check for an AAAA record, and if found, use that instead.

The assumption is that no A record will exist in the IPv6 environment,
which seems to hold for now.

This should enable the container runtime to resolve internal image
registry names for image pulls.

(cherry picked from commit 8d68ad43cd6d0d736e26e9f0256549c75a911249)